# Conflicts:
#	src/main/java/org/example/javabackend1/Room/RoomController.java
#	src/main/java/org/example/javabackend1/Room/RoomService.java
#	src/main/resources/templates/rooms/new.html
#	src/test/java/org/example/javabackend1/Booking/BookingServiceTest.java

+++ src/main/java/org/example/javabackend1/Customer/CustomerService.java
@@ -1,5 +1,6 @@
 package org.example.javabackend1.Customer;
 
+import jakarta.transaction.Transactional;
 import org.example.javabackend1.Exceptions.CustomerException;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +15,16 @@ public class CustomerService {
         this.customerRepository = customerRepository;
     }
 
+    public CustomerResponseDTO create(CustomerCreateDTO createDTO) {
+        if (customerRepository.findByEmail(createDTO.getEmail()).isPresent()) {
+            throw new CustomerException("Email already in use");
+        }
+
+        CustomerEntity customer = new CustomerEntity();
+
+        return createCustomerResponseDTO(createDTO, customer);
+    }
+
     public CustomerResponseDTO update(Long id, CustomerCreateDTO createDTO) {
         CustomerEntity customer = customerRepository.findById(id)
                 .orElseThrow(
@@ -23,43 +34,18 @@ public class CustomerService {
         return createCustomerResponseDTO(createDTO, customer);
     }
 
+    @Transactional
     public void delete(Long id) {
         CustomerEntity customer = customerRepository.findById(id)
                 .orElseThrow(
                         () -> new CustomerException("User with id " + String.valueOf(id) + " does not exist")
                 );
 
+        if (!customer.getBookings().isEmpty()) {
+            throw new CustomerException("Cannot delete customer with active bookings");
+        }
+
         customerRepository.delete(customer);
-    }
-
-    public CustomerResponseDTO findById(Long id) {
-        CustomerEntity customer = customerRepository.findById(id)
-                .orElseThrow(
-                        () -> new CustomerException("User with id " + String.valueOf(id) + " does not exist")
-                );
-
-        return toResponse(customer);
-    }
-
-    public List<CustomerResponseDTO> findAll() {
-        List<CustomerEntity> customers = customerRepository.findAll();
-
-        List<CustomerResponseDTO> responseBookings = new ArrayList<>();
-        for (CustomerEntity customer : customers) {
-            responseBookings.add(toResponse(customer));
-        }
-
-        return responseBookings;
-    }
-
-    public CustomerResponseDTO create(CustomerCreateDTO createDTO) {
-        if (customerRepository.findByEmail(createDTO.getEmail()).isPresent()) {
-            throw new CustomerException("Email already in use");
-        }
-
-        CustomerEntity customer = new CustomerEntity();
-
-        return createCustomerResponseDTO(createDTO, customer);
     }
 
     private CustomerResponseDTO createCustomerResponseDTO(CustomerCreateDTO createDTO, CustomerEntity customer) {
@@ -84,5 +70,25 @@ public class CustomerService {
         response.setPhone(customer.getPhone());
 
         return response;
+    }
+
+    public CustomerResponseDTO findById(Long id) {
+        CustomerEntity customer = customerRepository.findById(id)
+                .orElseThrow(
+                        () -> new CustomerException("User with id " + String.valueOf(id) + " does not exist")
+                );
+
+        return toResponse(customer);
+    }
+
+    public List<CustomerResponseDTO> findAll() {
+        List<CustomerEntity> customers = customerRepository.findAll();
+
+        List<CustomerResponseDTO> responseBookings = new ArrayList<>();
+        for (CustomerEntity customer : customers) {
+            responseBookings.add(toResponse(customer));
+        }
+
+        return responseBookings;
     }
 }
